split out the accumulator implementations, make them concurrent, add tests
