+++ core/src/main/scala/com/raphtory/api/analysis/graphstate/GraphStateImplementation.scala
@@ -2,238 +2,8 @@ package com.raphtory.api.analysis.graphstate
 
 import com.raphtory.utils.Bounded
 
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.concurrent.atomic.AtomicReference
-import scala.collection.Searching.Found
-import scala.collection.Searching.InsertionPoint
-import scala.collection.MapView
 import scala.collection.mutable
 
-private trait AccumulatorImplementation[@specialized(Int) -S, T] extends Accumulator[S, T] {
-  def currentValue: T
-  def merge(other: T): Unit
-  def reset(): Unit
-}
-
-private class SimpleAccumulatorImplementation[T](
-    initialValue: T,
-    retainState: Boolean = false,
-    op: (T, T) => T
-) extends AccumulatorImplementation[T, T] {
-  var currentValue: T = initialValue
-  var value: T        = initialValue
-
-  def +=(newValue: T): Unit =
-    this.synchronized(this.currentValue = op(currentValue, newValue))
-
-  def reset(): Unit = {
-    if (retainState)
-      value = op(value, currentValue)
-    else
-      value = currentValue
-    currentValue = initialValue
-  }
-
-  override def merge(other: T): Unit = this += other
-}
-import scala.jdk.FunctionConverters._
-
-private class ConcurrentAccumulatorImpl[T](initialValue: T, retainState: Boolean = false, op: (T, T) => T)
-        extends AccumulatorImplementation[T, T] {
-  val currentValueA            = new AtomicReference[T](initialValue)
-  val valueA                   = new AtomicReference[T](initialValue)
-  override def currentValue: T = currentValueA.get()
-
-  override def merge(other: T): Unit = this.+=(other)
-
-  override def reset(): Unit = {
-    // FIXME: possibly concurrency issues, we'll get back to this
-    if (retainState)
-      valueA.accumulateAndGet(currentValue, op.asJava)
-    else
-      valueA.set(currentValue)
-    currentValueA.set(initialValue)
-  }
-
-  /** Get last accumulated value */
-  override def value: T = valueA.get()
-
-  /** Add new value to accumulator
-    *
-    * @param newValue Value to add
-    */
-  override def +=(newValue: T): Unit =
-    this.currentValueA.accumulateAndGet(newValue, op.asJava)
-}
-
-private class IntAccumulatorImpl(initialValue: Int, retainState: Boolean = false, op: (Int, Int) => Int)
-        extends AccumulatorImplementation[Int, Int] {
-  private val currentValueA: AtomicInteger = new AtomicInteger(initialValue)
-  private val valueA: AtomicInteger        = new AtomicInteger(initialValue)
-  override def currentValue: Int           = currentValueA.get()
-
-  override def merge(other: Int): Unit = this += other
-
-  override def reset(): Unit = {
-    // FIXME: possibly concurrency issues, we'll get back to this
-    if (retainState)
-      valueA.accumulateAndGet(currentValue, op.asJava)
-    else
-      valueA.set(currentValue)
-    currentValueA.set(initialValue)
-  }
-
-  /** Get last accumulated value */
-  override def value: Int = valueA.get()
-
-  /** Add new value to accumulator
-    *
-    * @param newValue Value to add
-    */
-  override def +=(newValue: Int): Unit =
-    currentValueA.accumulateAndGet(newValue, op.asJava)
-}
-
-private object AccumulatorImplementation {
-
-  def apply[T](initialValue: T, retainState: Boolean = false, op: (T, T) => T) =
-    new SimpleAccumulatorImplementation[T](initialValue, retainState, op)
-
-  def concurrentInt(initialValue: Int, retainState: Boolean = false, op: (Int, Int) => Int) =
-    new IntAccumulatorImpl(initialValue = initialValue, retainState = retainState, op = op)
-}
-
-import scala.math.Numeric.Implicits.infixNumericOps
-
-private class CounterImplementation[T]() extends Counter[T] {
-  var totalCount: Int             = 0
-  var counts: mutable.Map[T, Int] = mutable.Map[T, Int]()
-
-  override def getCounts: MapView[T, Int]      = counts.view
-  override def largest: (T, Int)               = counts.maxBy(_._2)
-  override def largest(k: Int): List[(T, Int)] = counts.toList.sortBy(_._2).takeRight(k.min(counts.size))
-
-  def +=(newValue: (T, Int)): Unit =
-    this.synchronized {
-      if (counts.contains(newValue._1))
-        counts += (newValue._1 -> (counts(newValue._1) + newValue._2))
-      else
-        counts += (newValue._1 -> newValue._2)
-    }
-
-}
-
-private object CounterImplementation {
-  def apply[T]() = new CounterImplementation[T]()
-}
-
-private class CounterAccumulatorImplementation[T](retainState: Boolean)
-        extends AccumulatorImplementation[(T, Int), Counter[T]] {
-  var value: Counter[T]                      = CounterImplementation()
-  var currentValue: CounterImplementation[T] = CounterImplementation()
-
-  override def merge(other: Counter[T]): Unit = {
-    currentValue.totalCount += other.totalCount
-    val (map1, map2) = (currentValue.counts, other.getCounts)
-    val newCounts    = map1 ++ map2.map { case (k, v) => k -> (v + map1.getOrElse(k, 0)) }
-    currentValue.counts = newCounts
-  }
-
-  override def reset(): Unit = {
-    if (retainState)
-      merge(value)
-    val tmp = value.asInstanceOf[CounterImplementation[T]]
-    value = currentValue
-    currentValue = tmp
-    currentValue.totalCount = 0
-  }
-
-  /** Add new value to accumulator
-    *
-    * @param newValue Value to add
-    */
-  override def +=(newValue: (T, Int)): Unit = currentValue += newValue
-}
-
-private object CounterAccumulatorImplementation {
-  def apply[T](retainState: Boolean) = new CounterAccumulatorImplementation[T](retainState)
-}
-
-private class HistogramImplementation[T: Numeric](
-    val noBins: Int,
-    override val minValue: T,
-    override val maxValue: T
-) extends Histogram[T] {
-  val bins: Array[Int] = Array.fill(noBins)(0)
-  var totalCount: Int  = 0
-
-  override def getBins: Array[Int] = bins
-
-  override def cumSum: Array[Int] = bins.scanLeft(0)(_ + _)
-
-  override def quantile(quantile: Float): Float = {
-    val index = cumSum.search((totalCount * quantile).floor.toInt) match {
-      case InsertionPoint(insertionPoint) => insertionPoint - 1
-      case Found(foundIndex)              => foundIndex
-    }
-    minValue.toFloat + ((maxValue - minValue).toFloat / noBins) * index
-  }
-
-  override def getBin(value: T): Int =
-    (noBins * (value - minValue).toFloat / (maxValue - minValue).toFloat).floor.toInt
-
-  override def getBinCount(value: T): Int = bins(getBin(value))
-
-  def +=(value: T): Unit =
-    this.synchronized {
-      val bin = getBin(value).min(noBins - 1)
-      bins(bin) += 1
-      totalCount += 1
-    }
-}
-
-private object HistogramImplementation {
-
-  def apply[T: Numeric](noBins: Int, minValue: T, maxValue: T) =
-    new HistogramImplementation[T](noBins, minValue, maxValue)
-}
-
-private class HistogramAccumulatorImplementation[T: Numeric](
-    noBins: Int,
-    minValue: T,
-    maxValue: T,
-    retainState: Boolean
-) extends AccumulatorImplementation[T, Histogram[T]] {
-
-  var value: Histogram[T]                      = HistogramImplementation(noBins, minValue, maxValue)
-  var currentValue: HistogramImplementation[T] = HistogramImplementation(noBins, minValue, maxValue)
-
-  override def +=(value: T): Unit = currentValue += value
-
-  override def merge(other: Histogram[T]): Unit = {
-    currentValue.totalCount += other.totalCount
-    val otherBins = other.getBins
-    for (bindex <- currentValue.bins.indices)
-      currentValue.bins(bindex) += otherBins(bindex)
-  }
-
-  override def reset(): Unit = {
-    if (retainState)
-      merge(value)
-    val tmp = value.asInstanceOf[HistogramImplementation[T]]
-    value = currentValue
-    currentValue = tmp
-    currentValue.totalCount = 0
-    for (i <- currentValue.bins.indices)
-      currentValue.bins(i) = 0
-  }
-}
-
-private object HistogramAccumulatorImplementation {
-
-  def apply[T: Numeric](noBins: Int, minValue: T, maxValue: T, retainState: Boolean) =
-    new HistogramAccumulatorImplementation[T](noBins, minValue, maxValue, retainState)
-}
 
 private[raphtory] class GraphStateImplementation(override val nodeCount: Int) extends GraphState {
   private val accumulatorState = mutable.Map.empty[String, AccumulatorImplementation[Any, Any]]
